+++ packages/language-server/index.ts
@@ -1,4 +1,18 @@
-import * as ts from 'typescript';
 import { startServer } from './lib/server';
 
-startServer(ts);
+if (process.argv.includes('--version')) {
+	console.log(require('../package.json').version);
+}
+else {
+	let ts;
+	for (const arg of process.argv) {
+		if (arg.startsWith('--tsdk=')) {
+			const tsdk = arg.substring('--tsdk='.length);
+			const tsPath = require.resolve('./typescript.js', { paths: [tsdk] });
+			ts = require(tsPath);
+			break;
+		}
+	}
+	ts ??= require('typescript');
+	startServer(ts);
+}
